Refactored the InfoAlert component
