+++ src/components/Alert/InfoAlert.js
@@ -9,15 +9,18 @@ const useStyles = makeStyles({
     maxHeight: 300,
     overflowY: 'scroll',
   },
+  line: {
+    marginBottom: 10,
+  },
 });
 
 const InfoAlert = ({props}) => {
   const classes = useStyles();
   const state = useSelector(state => state);
 
-  if (state.alert.open) {
-    let text = state.alert.info.split('\n').map((line, i) => {
-      return <p key={i}>{line}</p>
+  if (state.infoAlert.open) {
+    let text = state.infoAlert.info.split('\n').map((line, i) => {
+      return <div className={classes.line} key={i}>{line}</div>
     });
     return (
       <Alert className={classes.infoAlert} severity="info">
