+++ src/reducers/infoAlertReducer.js
